Drop our own RTAX_INITCWND support, in favor of upstreams.
A patch to support hoplimit and initcwnd was submitted as Debian bug #221893.
Since then upstream seems to have implemented initcwnd, so the code was
duplicated in the debian package and the contributed patch in #221893 had a
cut'n'paste error (RTAX_HOPLIMIT vs RTAX_INITCWND).
This change simply drops our own initcwnd changes, while keeping the
hoplimit parts (for now).