Drop a couple of manpages now provided by upstream, and create some links.
Drop lnstat.8 and ss.8 from debian/man/ which are identical to the ones
upstream ships.
Leave debian/man/rtmon.8 in for now, since it differs and we need to figure
out if upstreams is newer or if our changes should be merged upstream.
Create manpage symlinks for compat command symlinks in debian/iproute.links
as we don't use "make install"....