Split out direct source edit for disabling build of ifstat.
Putting this as a dpatch for now, since this is probably something
that should be looked over again in a future cleanup.
Why are we disabling ifstat anyway?
Possible reason is to not conflict with the package ifstat, but
is the iproute2 ifstat not really useful at all?
If someone investigates this, please atleast update the dpatch
description.

Instead of patching the Makefile, we could just not install the file.
This would add some minimal extra build time for building something we
don't use anyway though.