Don't ship useless headers in iproute-dev (Closes: #467557)
The only useful thing in iproute-dev seems to be libnetlink.
Don't ship lots of internal headers which isn't useful for anyone else.
Also put the libnetlink.h in /usr/include, rather then /usr/include/iproute,
to have it in the same place as other distributions.

(build-rdepends only lists collectd as build-depending on iproute-dev, and
it seems to look for libnetlink.h in /usr/include{,/iproute,/linux} so it
should be safe to move it.)