Finally drop the patch adding the wrr qdisc scheduler to tc
The wrr qdisc support was never integrated in an upstream
kernel release and patches exists only for older kernels
which are all to be considered legacy nowadays.

The wrr patch for iproute does no longer apply against
the latest upstream iproute2 release (and it's caused
enough wasted time for me to not even consider making
it apply which probably wouldn't be a huge task).

If someone wants to revive wrr, please remember that iproute
is modular and that you can build wrr as a plugin module
for tc by putting your shared object file in /usr/lib/tc ...
See e.g. the atm qdisc.... so please package it separately
and maintain it yourself if you want to revive it!