Mark iproute as Multi-Arch: foreign.
Let us for now assume that no other packages ever provide additional
plugins. That means that the complete interface the iproute package is
architecture independent. So if any package ever attempts to ship
additional plugins, you need to revert the changes below.

If the extension mechanism is ever used, things get more complex. In
theory the package should be marked as "Multi-Arch: allowed" (i.e. it
provides both an architecture dependent and independent interface). It
requires ":any" dependencies which currently do not work. Thus this
option is currently unavailable from a practical point of view.

(Closes: #676175)