Generate recommends for optional module dependencies
Use dh_shlibdeps to dynamically generate a substvars
field with dependencies of modules which are optional
and we don't want to have a hard dependency on.

This should hopefully make it more visible that
iproute needs to be rebuilt whenever libxtables
bumps soname, to avoid breaking the tc xt module.