iproute2.postinst: use setcap -r instead of empty set
Calling setcap with an empty strings does not remove the xattr, but sets
it to an empty set, which breaks programs with ambient caps forking and
exec'ing ip.
Use setcap -r instead to remove the xattr if the user chooses to do so.

Thanks to Mantas Mikulėnas for spotting the error!