Make autopkgtests cross-test-friendly
In Ubuntu, we are in the process of moving the i386 architecture to a
compatibility-only layer on amd64, and therefore we are also moving our
autopkgtest infrastructure to test i386 binaries in a cross-environment.

This requires changes to some tests so that they are cross-aware and can do
the right thing.

The iproute2 tests currently fail in this environment, because they include
a build test that does not invoke the toolchain in a cross-aware manner and
because they try to install cross versions of base packages when they should
use the available host versions.  I've verified that the attached patch lets
the tests successfully build (and run) i386 tests on an amd64 host.

Closes: #949566