found a place to store the queue-sendoff timestamp
… however, it’s just an unsigned long; we’ll need an u64 for the
whole timestamp, which is true only on LP64 systems, or we could
store a 32-bit unsigned delta from cb->ts_enq which would limit
the one-way delay (including extralatency and queue delay) to a
little over 4 seconds though *and* would require error handling.

We could calculate with:
- up to 5*8=40 ms for retransmissions
  ⚠ HOWEVER we don’t know when we’ll be called by the kernel…
- up to 500+200=700 ms for qdelay plus some buffer
  (as all packets older than 500 ms are dropped every 200 ms)
- extralatency currently is limited to 2000 ms

This gives a total of 2800 ms plus buffers which might suffice.

It would complicate the code massively though, and we currently
good as require a 64-bit system already anyway, and this really
would already perform really badly on 32-bit systems, which makes
a good argument for requiring a 64-bit system now. This executive
decision may need revisiting but is good for now.